[CHORES] move CreateAssignment application from its own folder to another CreateAssignment folder
The folder change is done so that create and grade can be under one big folder umbrella
