+++ components/features/Assignment/Create/CreateAssignment.tsx
@@ -1,7 +1,7 @@
 import { Backdrop } from "@components/shared/layouts/Backdrop";
 import { MultistepIndex } from "@components/shared/layouts/MultistepAssignment";
 import React from "react";
-import { AssignmentForm } from "./AssignmentForm";
+import { AssignmentForm } from "../AssignmentForm";
 import styles from "./CreateAssignment.module.css";
 
 const CreateAssignment = () => {
